fix(gradle): pass extra env
