+++ lib/manager/gradle/index.ts
@@ -1,13 +1,11 @@
 import { Stats } from 'fs';
-import * as os from 'os';
-import { chmod, stat } from 'fs-extra';
+import { stat } from 'fs-extra';
 import upath from 'upath';
 import { LANGUAGE_JAVA } from '../../constants/languages';
 import * as datasourceMaven from '../../datasource/maven';
 import { logger } from '../../logger';
 import { ExternalHostError } from '../../types/errors/external-host-error';
 import { ExecOptions, exec } from '../../util/exec';
-import { BinarySource } from '../../util/exec/common';
 import { readLocalFile } from '../../util/fs';
 import * as gradleVersioning from '../../versioning/gradle';
 import {
@@ -26,43 +24,11 @@ import {
   createRenovateGradlePlugin,
   extractDependenciesFromUpdatesReport,
 } from './gradle-updates-report';
+import { extraEnv, gradleWrapperFileName, prepareGradleCommand } from './utils';
 
 export const GRADLE_DEPENDENCY_REPORT_OPTIONS =
   '--init-script renovate-plugin.gradle renovate';
 const TIMEOUT_CODE = 143;
-
-export function gradleWrapperFileName(config: ExtractConfig): string {
-  if (
-    os.platform() === 'win32' &&
-    config?.binarySource !== BinarySource.Docker
-  ) {
-    return 'gradlew.bat';
-  }
-  return './gradlew';
-}
-
-export async function prepareGradleCommand(
-  gradlewName: string,
-  cwd: string,
-  gradlew: Stats | null,
-  args: string | null
-): Promise<string> {
-  /* eslint-disable no-bitwise */
-  // istanbul ignore if
-  if (gradlew?.isFile() === true) {
-    // if the file is not executable by others
-    if ((gradlew.mode & 0o1) === 0) {
-      // add the execution permission to the owner, group and others
-      await chmod(upath.join(cwd, gradlewName), gradlew.mode | 0o111);
-    }
-    if (args === null) {
-      return gradlewName;
-    }
-    return `${gradlewName} ${args}`;
-  }
-  /* eslint-enable no-bitwise */
-  return null;
-}
 
 async function prepareGradleCommandFallback(
   gradlewName: string,
@@ -100,6 +66,7 @@ export async function executeGradle(
     docker: {
       image: 'renovate/gradle',
     },
+    extraEnv,
   };
   try {
     logger.debug({ cmd }, 'Start gradle command');
